refactor(capabilities): migrate the rpc server cap to the struct-hosted actor macro
Move the rpc server `impl NativeActor` into runtime.rs under `#[runtime]`
and put `#[actor(singleton)]` on the `RpcServerCapability` struct (#2386).
